+++ src/solver/types.jl
@@ -83,6 +83,22 @@ struct QuadratureCloud{SGQ <: AbstractSGSamplingType} <: AbstractCloudModel
 end
 
 
+"""
+    MLCloud
+
+Compute the cloud fraction using a machine learning model. Continue to use
+quadrature sampling for sub-grid variability of q_liq, q_ice.
+"""
+struct MLCloud{SGQ <: AbstractSGSamplingType, M} <: AbstractCloudModel
+    SG_quad::SGQ
+    model::M
+end
+
+function MLCloud_constructor(SG_quad, model)
+    static_model = Adapt.adapt_structure(SA.SArray, model)
+    return MLCloud{typeof(SG_quad), typeof(static_model)}(SG_quad, static_model)
+end
+
 abstract type AbstractSST end
 struct ZonallySymmetricSST <: AbstractSST end
 struct RCEMIPIISST <: AbstractSST end
@@ -97,70 +113,6 @@ end
 struct RCEMIPIIInsolation <: AbstractInsolation end
 struct GCMDrivenInsolation <: AbstractInsolation end
 struct ExternalTVInsolation <: AbstractInsolation end
-
-"""
-    AbstractOzone
-
-Describe how ozone concentration should be set.
-"""
-abstract type AbstractOzone end
-
-"""
-    IdealizedOzone
-
-Implement a static (not varying in time) idealized ozone profile as described by
-`idealized_ozone`.
-"""
-struct IdealizedOzone <: AbstractOzone end
-
-"""
-    PrescribedOzone
-
-Implement a time-varying ozone profile as read from disk.
-
-The CMIP6 forcing dataset is used. For production runs, you should acquire the
-high-resolution, multi-year `ozone_concentrations` artifact. If this is not available, a low
-resolution, single-year version will be used.
-
-Refer to ClimaArtifacts for more information on how to obtain the artifact.
-"""
-struct PrescribedOzone <: AbstractOzone end
-
-"""
-    AbstractCO2
-
-Describe how CO2 concentration should be set.
-"""
-abstract type AbstractCO2 end
-
-"""
-    FixedCO2
-
-Implement a static CO2 profile as read from disk.
-
-The data used is the one distributed with `RRTMGP.jl`.
-
-By default, this is 397.547 parts per million.
-
-This is the volume mixing ratio.
-"""
-struct FixedCO2{FT} <: AbstractCO2
-    value::FT
-
-    function FixedCO2(; FT = Float64, value = FT(397.547e-6))
-        return new{FT}(value)
-    end
-end
-
-"""
-    MuanaLoaCO2
-
-Implement a time-varying CO2 profile as read from disk.
-
-The data from the Mauna Loa CO2 measurements is used. It is a assumed that the
-concentration is constant.
-"""
-struct MaunaLoaCO2 <: AbstractCO2 end
 
 """
     AbstractCloudInRadiation
@@ -651,10 +603,8 @@ end
 
 Groups radiation-related models and types.
 """
-Base.@kwdef struct AtmosRadiation{RM, OZ, CO2, IN}
+Base.@kwdef struct AtmosRadiation{RM, IN}
     radiation_mode::RM = nothing
-    ozone::OZ = nothing
-    co2::CO2 = nothing
     insolation::IN = nothing
 end
 
@@ -857,8 +807,6 @@ model = AtmosModel(;
     moisture_model = EquilMoistModel(),
     microphysics_model = Microphysics0Moment(),
     radiation_mode = RRTMGPI.AllSkyRadiation(),
-    ozone = IdealizedOzone(),
-    co2 = FixedCO2()
 )
 ```
 
@@ -893,8 +841,6 @@ Internal testing and calibration components for single-column setups:
   - Global radiation: RRTMGPI.ClearSkyRadiation(), RRTMGPI.AllSkyRadiation()
   - Atmospheric forcing: HeldSuarezForcing() (for idealized dynamics)
   - SCM-specific: RadiationDYCOMS(), RadiationISDAC(), RadiationTRMM_LBA()
-- `ozone`: IdealizedOzone(), PrescribedOzone()
-- `co2`: FixedCO2(), MaunaLoaCO2()
 - `insolation`: IdealizedInsolation(), TimeVaryingInsolation(), etc.
 
 ## AtmosTurbconv
@@ -1124,8 +1070,6 @@ function EquilMoistAtmosModel(; kwargs...)
         surface_model = PrescribedSST(),
         sfc_temperature = ZonallySymmetricSST(),
         insolation = IdealizedInsolation(),
-        ozone = IdealizedOzone(),
-        co2 = FixedCO2(),
     )
     return AtmosModel(; defaults..., kwargs...)
 end
@@ -1144,8 +1088,6 @@ function NonEquilMoistAtmosModel(; kwargs...)
         surface_model = PrescribedSST(),
         sfc_temperature = ZonallySymmetricSST(),
         insolation = IdealizedInsolation(),
-        ozone = IdealizedOzone(),
-        co2 = FixedCO2(),
     )
     return AtmosModel(; defaults..., kwargs...)
 end